% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Autopopulates a New Analysis with Common Files}
\usage{
create_folders(list_of_dirs = c("R", "data", "results"))
}
\arguments{
\item{list_of_dirs}{A character vector of the names of folders to be created.
Defaults to c("R", "data", "results")}
}
\value{
Nothing explicit; this function is called for its side effects (action on files)
}
\description{
This function sets up the basic directory structure I use in most of my analyses.
By default, this will produce folders for R/, data/, and results/.
}
\examples{
create_folders()
}
