% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldersetup.R
\docType{package}
\name{foldersetup}
\alias{foldersetup}
\alias{foldersetup-package}
\title{Basic Compilation of RR Scripts}
\description{
This package serves to collect scripts and templates for producing reports in a more reproducible fashion
}
